pyproject.toml
[project]
name = "ultrathink"
version = "0.1.0"
description = "Self-Improving Development Framework with AI assistance"
authors = [
    {name = "PerryHMartin",email = "134838634+PerryHMartin@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic (>=0.69.0,<0.70.0)",
    "google-cloud-aiplatform (>=1.120.0,<2.0.0)",
    "boto3 (>=1.40.50,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "tree-sitter (>=0.25.2,<0.26.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "mcp (>=1.0.0,<2.0.0)"
]

[project.scripts]
ultrathink = "ultrathink.cli:run"

[tool.poetry]
packages = [{include = "ultrathink", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "ruff (>=0.14.0,<0.15.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
